+++ ru.taximaxim.codekeeper.ui/src/ru/taximaxim/codekeeper/ui/UIConsts.java
@@ -195,10 +195,6 @@ public interface UIConsts {
         String LAST_DB_STORE_CHANGE_TIME = "prefLastDbStoreChangeTime"; //$NON-NLS-1$
     }
 
-    interface CONN_TYPE_PREF {
-        String LAST_CONN_TYPE_CHANGE_TIME = "prefLastConnTypeChangeTime"; //$NON-NLS-1$
-    }
-
     interface PROJ_PATH {
         String MIGRATION_DIR = "MIGRATION"; //$NON-NLS-1$
     }